refactor distutil command subclasses
