+++ setup.py
@@ -27,9 +27,10 @@ setup(
     author='the scikit-build team',
     author_email='scikit-build@googlegroups.com',
     url='https://github.com/scikit-build/scikit-build',
-    packages=['skbuild', 'skbuild.platform_specifics'],
+    packages=['skbuild', 'skbuild.platform_specifics', 'skbuild.command'],
     package_dir={'skbuild': 'skbuild',
-                 'skbuild.platform_specifics': 'skbuild/platform_specifics'},
+                 'skbuild.platform_specifics': 'skbuild/platform_specifics',
+                 'skbuild.command': 'skbuild/command'},
     include_package_data=True,
     install_requires=requirements,
     license="MIT",
